Add files via upload
Added simple version of tool creation experiment script. This includes a function construct_policies() which creates a restricted number of all policies and this list can then be passed into the agent class. The function run_AIF_loop() runs the simulation in a single function. So far this function does not include any plotting functions, it only renders the environment images for you.